Refactored FilterBinder to reuse and added support for prop op prop in clauses
