+++ OData/src/System.Web.OData/OData/Query/ApplyQueryOption.cs
@@ -186,7 +186,7 @@ namespace System.Web.OData.Query
                 else
                 {
 
-                    var binder = new AggregationBinder(querySettings, assembliesResolver, elementType, Context.Model, transformation);
+                    var binder = new AggregationBinder(updatedSettings, assembliesResolver, elementType, Context.Model, transformation);
                     query =  binder.Bind(query);
                     elementType = binder.ResultType;
                 }
