+++ OData/src/System.Web.OData/System.Web.OData.csproj
@@ -129,6 +129,7 @@
     <Compile Include="OData\Query\Expressions\AggregationBinder.cs" />
     <Compile Include="OData\Query\ApplyQueryOption.cs" />
     <Compile Include="OData\Query\Expressions\DynamicTypeWrapper.cs" />
+    <Compile Include="OData\Query\Expressions\ExpressionBinderBase.cs" />
     <Compile Include="OData\Query\Expressions\IdentityPropertyMapper.cs" />
     <Compile Include="OData\Query\Expressions\TypeDefinition.cs" />
     <Compile Include="OData\Query\Expressions\TypeProvider.cs" />
